gamb-back/requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.12.3
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
constantly==23.10.4
cryptography==44.0.1
daphne==4.1.2
Django==4.2.15
django-bootstrap-v5==1.0.11
django-cors-headers==4.6.0
django-filter==2.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
msgpack==1.1.0
packaging==24.2
pillow==10.4.0
psycopg==3.2.3
psycopg-binary==3.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.0.0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.35.1
rpds-py==0.21.0
service-identity==24.2.0
setuptools==75.8.0
soupsieve==2.5
sqlparse==0.5.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
zope.interface==7.2
